requirements.txt
anyjson==0.3.3
celery==4.1.0
defusedxml==0.4.1
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.6
django-filter==0.13.0
django-registration==2.2
djangorestframework==3.3.3
lxml==3.6.0
psycopg2==2.6.1
requests==2.10.0
static3==0.7.0
waitress==0.9.0
wheel==0.29.0
